--- src/main/java/pro/sky/Galculator/CalculatorInterface.java
@@ -1,10 +1,0 @@
-package pro.sky.Galculator;
-
-public interface CalculatorInterface {
-    String greet();
-    double add(double a, double b);
-    double subtract(double a, double b);
-    double multiply(double a, double b);
-    double divide(double a, double b);
-
-}
