+++ src/main/java/pro/sky/Galculator/CalculatorService.java
@@ -1,28 +1,10 @@
 package pro.sky.Galculator;
 
-import org.springframework.stereotype.Service;
+public interface CalculatorService {
+    String greet();
+    double add(double a, double b);
+    double subtract(double a, double b);
+    double multiply(double a, double b);
+    double divide(double a, double b);
 
-@Service
-public class CalculatorService implements CalculatorInterface{
-    public String greet() {
-        return "<big>Добро пожаловать в калькулятор</big>";
-    }
-
-    public double add(double a, double b) {
-        return a + b;
-    }
-
-    public double subtract(double a, double b) {
-        return a - b;
-    }
-
-
-    public double multiply(double a, double b) {
-        return a * b;
-    }
-
-
-    public double divide(double a, double b) {
-        return a / b;
-    }
 }
